TeamPJT/Study/requirements.txt
findspark==1.4.2
py4j==0.10.7
pyspark==2.4.6
python-twitter==3.5
kafka-python==2.0.2
